requirements.txt
Flask==2.2.2
MLB_StatsAPI==1.4.1
pytz==2021.3
requests==2.28.1
